+++ src/thumbor_video_engine/app.py
@@ -5,10 +5,38 @@ import tornado.gen as gen
 
 import thumbor.app
 from thumbor.handlers.imaging import ImagingHandler
+from thumbor.result_storages import ResultStorageResult
+from thumbor_video_engine.utils import is_animated, is_animated_gif
 
 
 class VideoEngineImagingHandler(ImagingHandler):
     def _override_write_results_to_client(self, results, content_type):
+        is_gif = content_type == 'image/gif'
+        is_webp = content_type == 'image/webp'
+        is_video = content_type.startswith('video/')
+        accepts_video = getattr(self.context.request, "accepts_video", False)
+        accepts_webp = self.context.request.accepts_webp
+        auto_gif_video = (
+            self.context.config.FFMPEG_GIF_AUTO_H264 or
+            self.context.config.FFMPEG_GIF_AUTO_H265)
+        auto_gif_webp = (
+            self.context.config.FFMPEG_GIF_AUTO_WEBP and
+            self.context.config.AUTO_WEBP)
+
+        # If the result is from result_storage, we need to determine whether
+        # we should Vary the result because of auto-gif-conversion settings
+        if isinstance(results, ResultStorageResult):
+            buf = results.buffer
+            if is_webp and is_animated(buf) and accepts_webp and auto_gif_webp:
+                self.context.request.should_vary = True
+            elif is_video and accepts_video and auto_gif_video:
+                self.context.request.should_vary = True
+            elif is_gif and is_animated_gif(buf):
+                if auto_gif_video and not accepts_video:
+                    self.context.request.should_vary = True
+                elif auto_gif_webp and not accepts_webp:
+                    self.context.request.should_vary = True
+
         # RequestParameters.filters is an empty list when none are in the url,
         # and ImagingHandler._write_results_to_client assumes that if
         # context.request.format is set then it came from the format filter.